Fixed gen_marks.m
gen_marks.m is a file intended to automatically find the f0 inside
complex audio files. This'll be later used for the PSOLA algorithm.
Currently the markings are found using the findpeaks function from
octave, however, in the future a more complex method should be used
for example, YIN, a fundamental frequency estimator for speech and
music.